+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * docs/guides/*: Added some basic guides on how to manage different resources with dependencies
 * docs/resources/subaccount.md: Add group_ids to example
 
+### Bug Fixes:
+* provider/resources/objectgroup.go: Fixed updates in place for `options` on apply
+
 ## 1.0.15
 
 ### Enhancements: